Remove redundant newline from "literal" output
The "literal" command was adding newlines to every line, but this
is not necessary since the input newlines are preserved.

This is a rather ironic bug given the name of the command ;)

NB: due to how gsl_read_file works, this change causes input lines with
CR line endings to be output as a single, long line.