Problem: displaying of $(...) difficult
It's either ugly if written as &#36;(...) (note that `...` can't be used as it
will show &#36; literally, not replace it with a dollar-sign), or it
causes gitdown to issue errors if written as `$(...)`.

Solution: use <tt>&#36;(...)</tt> trick

Apparently, this still replaces the HTML escape sequence with the
dollar-sign, but is still "invisible" to gitdown, so it won't issue any
errors about undefined variables.